+++ packages/solid-picture/src/index.tsx
@@ -1,3 +1,4 @@
 export { Img } from './Img'
 export { Source } from './Source'
 export { Picture } from './Picture'
+export type { MediaElementProps } from './types'